1. 视觉速度取int类型;   2. 分析找不到glAlg的bug解决方式;

+++ SMG_NothingIsAll/CortexAlgorithms/Vision/AIVisionAlgsModel.h
@@ -28,7 +28,7 @@
 //@property (assign,nonatomic) CGFloat radius;
 
 //speed
-@property (assign,nonatomic) CGFloat speed;
+@property (assign,nonatomic) NSInteger speed;
 
 //direction
 @property (assign,nonatomic) CGFloat direction;
